Add local path replacement for Polaris module
