+++ go.mod
@@ -2,6 +2,8 @@ module github.com/fairwindsops/polaris
 
 go 1.20
 
+replace github.com/fairwindsops/polaris => /Users/vvezani/fairwinds/polaris
+
 require (
 	github.com/AlecAivazis/survey/v2 v2.3.7
 	github.com/fairwindsops/controller-utils v0.3.0
