+++ pkg/mutation/mutate_test.go
@@ -25,14 +25,15 @@ var testCases = []struct {
 	mutated  string
 	patch    config.Mutation
 	message  string
-}{{
-	original: oldYaml,
-	patch: config.Mutation{
-		Op:    "add",
-		Value: "Denver",
-		Path:  "/pets/0/owners/*/location",
-	},
-	mutated: `pets:
+}{
+	{
+		original: oldYaml,
+		patch: config.Mutation{
+			Op:    "add",
+			Value: "Denver",
+			Path:  "/pets/0/owners/*/location",
+		},
+		mutated: `pets:
   - name: fido
     owners:
       - name: Alice
@@ -43,27 +44,27 @@ var testCases = []struct {
         location: Denver
   - name: scooby
 `,
-}, {
-	original: oldYaml,
-	patch: config.Mutation{
-		Op:   "remove",
-		Path: "/pets/0/owners/*/aliases",
-	},
-	mutated: `pets:
+	}, {
+		original: oldYaml,
+		patch: config.Mutation{
+			Op:   "remove",
+			Path: "/pets/0/owners/*/aliases",
+		},
+		mutated: `pets:
   - name: fido
     owners:
       - name: Alice
       - name: Bob
   - name: scooby
 `,
-}, {
-	original: oldYaml,
-	patch: config.Mutation{
-		Op:    "add",
-		Value: "rob",
-		Path:  "/pets/0/owners/*/aliases/-",
-	},
-	mutated: `pets:
+	}, {
+		original: oldYaml,
+		patch: config.Mutation{
+			Op:    "add",
+			Value: "rob",
+			Path:  "/pets/0/owners/*/aliases/-",
+		},
+		mutated: `pets:
   - name: fido
     owners:
       - name: Alice
@@ -75,23 +76,23 @@ var testCases = []struct {
           - rob
   - name: scooby
 `,
-}, {
-	original: `
+	}, {
+		original: `
 pets:
   - name: fido
 `,
-	patch: config.Mutation{
-		Op:    "add",
-		Value: "Alice",
-		Path:  "/pets/0/owners/0/name",
-	},
-	mutated: `pets:
+		patch: config.Mutation{
+			Op:    "add",
+			Value: "Alice",
+			Path:  "/pets/0/owners/0/name",
+		},
+		mutated: `pets:
   - name: fido
     owners:
       - name: Alice
 `,
-}, {
-	original: `
+	}, {
+		original: `
 obj:
   foo:
     bar:
@@ -99,54 +100,131 @@ obj:
       - b
   baz: quux
 `,
-	patch: config.Mutation{
-		Op: "replace",
-		Value: map[string]interface{}{
-			"bar": []string{"c", "d"},
+		patch: config.Mutation{
+			Op: "replace",
+			Value: map[string]interface{}{
+				"bar": []string{"c", "d"},
+			},
+			Path: "/obj/foo",
 		},
-		Path: "/obj/foo",
-	},
-	mutated: `obj:
+		mutated: `obj:
   foo:
     bar:
       - c
       - d
   baz: quux
 `,
-}, {
-	original: `
+	}, {
+		original: `
 foo: bar
 `,
-	patch: config.Mutation{
-		Op:      "replace",
-		Value:   "baz",
-		Path:    "/foo",
-		Comment: "# We set this to baz",
-	},
-	mutated: `
+		patch: config.Mutation{
+			Op:      "replace",
+			Value:   "baz",
+			Path:    "/foo",
+			Comment: "# We set this to baz",
+		},
+		mutated: `
 foo: baz # We set this to baz
 `,
-	message: "Expected a comment to appear",
-}, {
-	original: `
+		message: "Expected a comment to appear",
+	}, {
+		original: `
 foo: bar
 `,
-	patch: config.Mutation{
-		Op: "add",
-		Value: map[string]interface{}{
-			"baz": "quux",
+		patch: config.Mutation{
+			Op: "add",
+			Value: map[string]interface{}{
+				"baz": "quux",
+			},
+			Path:    "/extra",
+			Comment: "# These are extra things",
 		},
-		Path:    "/extra",
-		Comment: "# These are extra things",
-	},
-	mutated: `
+		mutated: `
 foo: bar
 extra:
   # These are extra things
   baz: quux
 `,
-	message: "Expected a comment to appear next to an object",
-}}
+		message: "Expected a comment to appear next to an object",
+	},
+	{
+		original: `
+foo: bar
+`,
+		patch: config.Mutation{
+			Op: "replace",
+			Value: map[string]interface{}{
+				"baz": "quux",
+			},
+			Path:    "/extra",
+			Comment: "# These are extra things",
+		},
+		mutated: `
+foo: bar
+extra:
+  # These are extra things
+  baz: quux
+`,
+		message: "Expected a comment to appear next to an object",
+	},
+	{
+		original: `
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: nginx-deployment
+  labels:
+    app: nginx
+spec:
+  template:
+    spec:
+      containers:
+        - name: nginx
+          image: nginx:1.18-alpine
+          ports:
+            - containerPort: 80
+`,
+		patch: config.Mutation{
+			Op: "add",
+			Value: map[string]any{
+				"limits": map[string]any{
+					"cpu":    "25m",
+					"memory": "100Mi",
+				},
+				"requests": map[string]any{
+					"cpu":    "25m",
+					"memory": "100Mi",
+				},
+			},
+			Path: "/spec/template/spec/containers/*/resources",
+		},
+		mutated: `
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: nginx-deployment
+  labels:
+    app: nginx
+spec:
+  template:
+    spec:
+      containers:
+        - name: nginx
+          image: nginx:1.18-alpine
+          ports:
+            - containerPort: 80
+          resources:
+            limits:
+              cpu: 25m
+              memory: 100Mi
+            requests:
+              cpu: 25m
+              memory: 100Mi
+`,
+		message: "Expected a comment to appear next to an object",
+	},
+}
 
 func TestApplyAllMutations(t *testing.T) {
 	for _, tc := range testCases {
